pyproject.toml
[project]
name = "lms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi>=0.127.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "imagekitio>=5.0.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.40.0",
]
